Win32 build support with GLEW, fixed test app

+++ include/cubicvr2/core/Timer.h
@@ -32,7 +32,7 @@ private:
 
 	bool paused_state;
 	bool lock_state;
-	float lock_rate;
+	unsigned long lock_rate;
 
 public:
